test: add SuperParkingBoy and not full parkingLots and Cars test

+++ src/main/java/com/afs/parkinglot/ParkingLot.java
@@ -42,8 +42,8 @@ public class ParkingLot {
         return capacity-parkedCars.size();
     }
 
-    public int getAvailableCapacityRatio() {
-        return getAvailableCapacity()/capacity;
+    public float getAvailableCapacityRatio() {
+        return (float) getAvailableCapacity() /capacity;
     }
 
 }
